+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Upgrade to Alchemy v1.4.0
 * Add support for `titleize` style parameter
 * The `fallback` property will not default to the `key` property anymore
+* Translation records with a negative weight will not be used when a fallback is given
 
 ## 0.6.8 (2024-01-22)